WIP refacto : explode v3 controller, fix warnings
